+++ src/support/defaults.js
@@ -6,5 +6,6 @@ const VERBOSITY = require(join(__appRoot, 'src/support/verbosity'));
 module.exports = {
   iterations: 1000,
   verbosity: VERBOSITY.NORMAL,
-  testdataMagnitude: 1000
+  testdataMagnitude: 1000,
+  precision: 4
 };